% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_coeftest.R
\name{tidy_coeftest}
\alias{tidy_coeftest}
\title{Prep models for visualization}
\usage{
tidy_coeftest(x, vcov. = NULL, model = NULL)
}
\arguments{
\item{x}{A model object that is compatible with \code{lmtest::coeftest()}.}

\item{vcov.}{A variance-covariance matrix for coefficient standard errors. If \code{NULL} (default) classical standard errors are produced. Will accept any variance-covariance matrix produced by alternative methods, for example via functions from the \code{{sandwhich}} package for robust and clustered standard errors.}

\item{model}{An optional model label. Must be a character string. If \code{NULL} (default), the label \code{"Model 1"} is applied.}
}
\value{
A tidy coefficient test tibble with eight columns: \code{term}, \code{estimate}, \code{std.error}, \code{statistic}, \code{p.value}, \code{conf.low}, \code{conf.high}, \code{model}, and \code{N}.
}
\description{
\code{tidy_coeftest()} takes a model object and converts
it into a tidy table of coefficients and their associated
standard errors, test statistics, etc. in a format that is
acceptable for use with \code{ggregtab()}.
}
